+++ src/Modules/Execution/Execution.Application/Features/Activities/Queries/ActivityQueries.cs
@@ -3,4 +3,4 @@ using MediatR;
 namespace Himapp.Execution.Application.Features.Activities.Queries;
 
 public sealed record GetAllActivitiesQuery : IRequest<IReadOnlyCollection<ActivityDto>>;
-public sealed record GetActivityByIdQuery(long Id) : IRequest<ActivityDto?>;
+public sealed record GetActivityByIdQuery(int Id) : IRequest<ActivityDto?>;
